# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Vacctrx Api",
    "version": "1.0.0",
    "description": "Rest Api's for Vacctrax Application."
  },
  "servers": [
    {
      "url": "https://vacctrax-backend.herokuapp.com/api",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
  "paths": {
    "/requestAppointment": {
      "post": {
        "tags": ["Appointment"],
        "summary": "Request an Appointment",
        "operationId": "requestAppointment",
        "requestBody": {
          "description": "Request Appointment Body",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestAppointmentDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Operation Successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Appointment"
                }
              }
            }
          },
          "400": {
            "description": "Invalid Json",
            "content": {

            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Address": {
        "type": "object",
        "required": [
          "streetLine1",
          "postalCode",
          "province",
          "city"
        ],
        "properties": {
          "streetLine1": {
            "type": "string",
            "example": "43, Test Street"
          },
          "streetLine2": {
            "type": "string",
            "example": "45, Test Street"
          },
          "postalCode": {
            "type": "string",
            "example": "J7J7J7"
          },
          "city": {
            "type": "string",
            "example": "Scarborough"
          }
        }
      },
      "Account": {
        "type": "object",
        "required": [
          "firstName",
          "lastName",
          "email",
          "password",
          "phone",
          "address",
          "type"
        ],
        "properties": {
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "email": {
            "type": "string",
            "example": "john@gmail.com"
          },
          "password": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "$ref": "#/components/schemas/Address"
          },
          "type": {
            "type": "string",
            "enum": ["PATIENT", "MEDICALADMIN", "HEALTHPRACTITIONER", "GOVERNMENT"],
            "description": "Account Type"
          }
        }
      },
      "Clinic": {
        "type": "object",
        "required": [
          "name",
          "address"
        ],
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "example": "Centennial Clinic"
          },
          "address": {
            "$ref": "#/components/schemas/Address"
          }
        }
      },
...
      "Patient": {
        "type": "object",
        "required": [
          "healthCardNo",
          "account"
        ],
        "properties": {
          "healthCardNo": {
            "type": "string"
          },
          "account": {
            "$ref": "#/components/schemas/Account"
          }
        }
      },
      "HealthPractitioner": {
        "type": "object",
        "required": [
          "account",
          "clinic"
        ],
        "properties": {
          "_id": {
            "type": "string"
          },
          "account": {
            "$ref": "#/components/schemas/Account"
          },
          "clinic": {
            "$ref": "#/components/schemas/Clinic"
          }
        }
      },
...
      "Appointment": {
        "type": "object",
        "required": [
          "reason",
          "preferredDate",
          "preferredTime",
          "type",
          "clinic",
          "patient"
        ],
        "properties": {
          "_id": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "preferredDate": {
            "type": "string",
            "format": "date-time"
          },
          "preferredTime": {
            "type": "string",
            "format": "date-time"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "type": {
            "type": "string",
            "enum": ["CONFIRMED", "REQUESTED", "CANCELLED", "COMPLETED"],
            "default": "REQUESTED"
          },
          "clinic": {
            "$ref": "#/components/schemas/Clinic"
          },
          "patient": {
            "$ref": "#/components/schemas/Patient"
          },
          "healthPractitioner": {
            "$ref": "#/components/schemas/HealthPractitioner"
          }
        }
      },
      "RequestAppointmentDTO": {
        "type": "object",
        "required": [
          "reason",
          "preferredDate",
          "preferredTime",
          "type",
          "clinic",
          "patient"
        ],
        "properties": {
          "reason": {
            "type": "string"
          },
          "preferredDate": {
            "type": "string",
            "format": "date-time"
          },
          "preferredTime": {
            "type": "string",
            "format": "date-time"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "vaccineDose": {
            "type": "number",
            "example": 1
          },
          "vaccine": {
            "type": "string",
            "example": "Pfizer"
          },
          "type": {
            "type": "string",
            "enum": ["CONFIRMED", "REQUESTED", "CANCELLED", "COMPLETED"],
            "default": "REQUESTED"
          },
          "clinicId": {
            "type": "string"
          },
          "patientId": {
            "type": "string"
          },
          "healthPractitionerId": {
            "type": "string"
          }
        }
      }
    }
  }
}